+++ src/schemathesis/lazy.py
@@ -1,12 +1,13 @@
 # pylint: disable=too-many-instance-attributes
 from inspect import signature
-from typing import Any, Callable, Dict, Optional, Union
+from typing import Any, Callable, Dict, Iterable, Optional, Union
 
 import attr
 import pytest
 from _pytest.fixtures import FixtureRequest
 from pytest_subtests import SubTests
 
+from .constants import DEFAULT_DATA_GENERATION_METHODS, DataGenerationMethod
 from .exceptions import InvalidSchema
 from .hooks import HookDispatcher, HookScope
 from .models import Endpoint
@@ -25,6 +26,7 @@ class LazySchema:
     hooks: HookDispatcher = attr.ib(factory=lambda: HookDispatcher(scope=HookScope.SCHEMA))  # pragma: no mutate
     validate_schema: bool = attr.ib(default=True)  # pragma: no mutate
     skip_deprecated_endpoints: bool = attr.ib(default=False)  # pragma: no mutate
+    data_generation_methods: Iterable[DataGenerationMethod] = attr.ib(default=DEFAULT_DATA_GENERATION_METHODS)
 
     def parametrize(  # pylint: disable=too-many-arguments
         self,
@@ -34,6 +36,7 @@ class LazySchema:
         operation_id: Optional[Filter] = NOT_SET,
         validate_schema: Union[bool, NotSet] = NOT_SET,
         skip_deprecated_endpoints: Union[bool, NotSet] = NOT_SET,
+        data_generation_methods: Union[Iterable[DataGenerationMethod], NotSet] = NOT_SET,
     ) -> Callable:
         if method is NOT_SET:
             method = self.method
@@ -43,6 +46,8 @@ class LazySchema:
             tag = self.tag
         if operation_id is NOT_SET:
             operation_id = self.operation_id
+        if data_generation_methods is NOT_SET:
+            data_generation_methods = self.data_generation_methods
 
         def wrapper(func: Callable) -> Callable:
             def test(request: FixtureRequest, subtests: SubTests) -> None:
@@ -60,6 +65,7 @@ class LazySchema:
                     test_function=func,
                     validate_schema=validate_schema,
                     skip_deprecated_endpoints=skip_deprecated_endpoints,
+                    data_generation_methods=data_generation_methods,
                 )
                 fixtures = get_fixtures(func, request)
                 # Changing the node id is required for better reporting - the method and endpoint will appear there
@@ -67,9 +73,9 @@ class LazySchema:
                 settings = getattr(test, "_hypothesis_internal_use_settings", None)
                 tests = list(schema.get_all_tests(func, settings))
                 request.session.testscollected += len(tests)
-                for _endpoint, sub_test in tests:
+                for _endpoint, data_generation_method, sub_test in tests:
                     actual_test = get_test(sub_test)
-                    subtests.item._nodeid = _get_node_name(node_id, _endpoint)
+                    subtests.item._nodeid = _get_node_name(node_id, _endpoint, data_generation_method)
                     run_subtest(_endpoint, fixtures, actual_test, subtests)
                 subtests.item._nodeid = node_id
 
@@ -93,9 +99,9 @@ def get_test(test: Union[Callable, InvalidSchema]) -> Callable:
     return test
 
 
-def _get_node_name(node_id: str, endpoint: Endpoint) -> str:
+def _get_node_name(node_id: str, endpoint: Endpoint, data_generation_method: DataGenerationMethod) -> str:
     """Make a test node name. For example: test_api[GET:/users]."""
-    return f"{node_id}[{endpoint.method.upper()}:{endpoint.full_path}]"
+    return f"{node_id}[{endpoint.method.upper()}:{endpoint.full_path}][{data_generation_method.as_short_name()}]"
 
 
 def run_subtest(endpoint: Endpoint, fixtures: Dict[str, Any], sub_test: Callable, subtests: SubTests) -> None:
@@ -116,6 +122,7 @@ def get_schema(
     hooks: HookDispatcher,
     validate_schema: Union[bool, NotSet] = NOT_SET,
     skip_deprecated_endpoints: Union[bool, NotSet] = NOT_SET,
+    data_generation_methods: Union[Iterable[DataGenerationMethod], NotSet] = NOT_SET,
 ) -> BaseSchema:
     """Loads a schema from the fixture."""
     # pylint: disable=too-many-arguments
@@ -131,6 +138,7 @@ def get_schema(
         hooks=hooks,
         validate_schema=validate_schema,
         skip_deprecated_endpoints=skip_deprecated_endpoints,
+        data_generation_methods=data_generation_methods,
     )